add few tests for react component
