+++ .npmignore
@@ -1,5 +1,6 @@
 node_modules
 build
+tests
 examples
 dist/*.html
 docs
