+++ .travis.yml
@@ -1,0 +1,8 @@
+language: node_js
+sudo: false
+node_js:
+  - "node"
+before_script:
+  - export DISPLAY=:99.0
+  - export TRAVIS=true
+  - sh -e /etc/init.d/xvfb start
