+++ examples/react-todomvc/index.html
@@ -7,6 +7,6 @@
     </div>
   </body>
   <!--<script src="node_modules/component-inspector/src/react.js"></script>-->
-  <script src="node_modules/component-inspector/dist/react.js"></script>
+  <script src="/static/inspector.js"></script>
   <script src="/static/bundle.js"></script>
 </html>
