+++ examples/react-todomvc/package.json
@@ -28,28 +28,28 @@
   },
   "homepage": "https://github.com/gaearon/redux#readme",
   "dependencies": {
-    "classnames": "^2.1.2",
-    "react": "^0.14.2",
-    "react-dom": "^0.14.2",
-    "redux": "^3.0.0",
-    "react-redux": "^3.0.1",
-    "todomvc-app-css": "^2.0.1"
+    "classnames": "2.1.2",
+    "react": "0.14.2",
+    "react-dom": "0.14.2",
+    "redux": "3.0.0",
+    "react-redux": "3.0.1",
+    "todomvc-app-css": "2.0.1"
   },
   "devDependencies": {
-    "babel-core": "^5.8.22",
-    "babel-loader": "^5.3.2",
-    "babel-plugin-react-display-name": "^2.0.0",
-    "babel-plugin-source-wrapper": "^1.1.1",
-    "chai": "^3.0.0",
+    "babel-core": "5.8.22",
+    "babel-loader": "5.3.2",
+    "babel-plugin-react-display-name": "2.0.0",
+    "babel-plugin-source-wrapper": "1.1.1",
+    "chai": "3.0.0",
     "component-inspector": "1.2.0",
-    "express-open-in-editor": "^1.0.0",
-    "mocha": "^2.2.5",
-    "node-libs-browser": "^0.5.2",
-    "open-in-editor": "^1.0.1",
-    "raw-loader": "^0.5.1",
-    "react-hot-loader": "^1.2.7",
-    "style-loader": "^0.12.3",
-    "webpack": "^1.12.2",
-    "webpack-dev-server": "^1.12.0"
+    "express-open-in-editor": "1.0.0",
+    "mocha": "2.2.5",
+    "node-libs-browser": "0.5.2",
+    "open-in-editor": "1.0.1",
+    "raw-loader": "0.5.1",
+    "react-hot-loader": "1.2.7",
+    "style-loader": "0.12.3",
+    "webpack": "1.12.2",
+    "webpack-dev-server": "1.12.0"
   }
 }
