+++ package.json
@@ -18,11 +18,30 @@
   "licenses": "MIT",
   "dependencies": {},
   "devDependencies": {
-    "basisjs": "^1.5.1",
-    "basisjs-tools": "^1.6.1"
+    "babel-core": "5.8.22",
+    "babel-loader": "5.3.2",
+    "babel-plugin-react-display-name": "2.0.0",
+    "babel-plugin-source-wrapper": "1.1.1",
+    "basisjs": "1.5.1",
+    "basisjs-tools": "1.6.1",
+    "chai": "3.5.0",
+    "classnames": "2.1.2",
+    "karma": "0.13.22",
+    "karma-chrome-launcher": "1.0.1",
+    "karma-firefox-launcher": "^1.0.0",
+    "karma-mocha": "1.0.1",
+    "karma-webpack": "1.7.0",
+    "mocha": "2.5.1",
+    "react": "0.14.2",
+    "react-dom": "0.14.2",
+    "react-hot-loader": "1.2.7",
+    "react-redux": "3.0.1",
+    "redux": "3.0.0",
+    "webpack": "1.13.1"
   },
   "scripts": {
     "build": "basis build --pack",
-    "prepublish": "npm run build"
+    "prepublish": "npm run build",
+    "test": "karma start"
   }
 }
